feat: deprecate old navbar (#494)
* feat: remove deprecated component

* feat: remove navbar

* feat: rename Navbar2 to Navbar

* fear: rename

* fix: not export old component

* fix: rename

* fix: wrong imports

* fix: prettier

* wip

* rename css

* rollback

BREAKING CHANGE: rename Navbar2 to Navbar, and removes old Navbar and UserInformation components
