--- src/components/UserInformationContainer/UserInformationContainer.css
@@ -1,260 +1,0 @@
-.dcl.new-user-menu {
-  display: flex;
-  position: relative;
-  text-align: left;
-  outline: none;
-  flex-direction: row;
-  gap: 15px;
-  align-items: center;
-}
-
-.dcl.new-user-menu .toggle,
-.dcl.new-user-menu .toggle .avatar-face {
-  width: 42px;
-  height: 42px;
-}
-
-.dcl.new-user-menu .menu .avatar-face {
-  background: rgba(var(--dark-raw), 0.3);
-}
-
-.dcl.new-user-menu .toggle {
-  cursor: pointer;
-}
-
-.dcl.new-user-menu .ui.button {
-  font-size: 12px;
-  line-height: 18px;
-  padding: 9px 14px;
-  min-width: 0;
-}
-
-.dcl.new-user-menu .menu {
-  position: absolute;
-  top: 56px;
-  right: 0;
-  z-index: 1;
-  background: var(--dropdown);
-  box-shadow: var(--shadow-2);
-  border-radius: 6px;
-  display: inline-block;
-  transition: transform 200ms ease-in-out, opacity 200ms ease-in-out;
-  opacity: 0;
-  user-select: none;
-  pointer-events: none;
-  transform: translate(0, -4px);
-  min-width: 270px;
-  overflow: hidden;
-}
-
-.dcl.new-user-menu .menu.clickable {
-  pointer-events: auto;
-}
-
-.dcl.new-user-menu .menu.open {
-  opacity: 1;
-  transform: translate(0, 0);
-  display: flex;
-  flex-direction: column;
-}
-
-.dcl.new-user-menu .info {
-  color: var(--text);
-  display: flex;
-  align-items: center;
-  padding: 16px;
-  justify-content: center;
-}
-
-.dcl.new-user-menu .info.clickable {
-  cursor: pointer;
-}
-
-.dcl.new-user-menu .dcl.new-account-wrapper .clickable {
-  cursor: pointer;
-}
-
-.dcl.new-user-menu .info.clickable:hover {
-  background-color: var(--dropdown-hover);
-}
-
-.dcl.new-user-menu .image {
-  margin-right: 10px;
-}
-
-.dcl.new-user-menu .name {
-  font-weight: 500;
-  font-size: 14px;
-  line-height: 20px;
-  letter-spacing: -0.19px;
-}
-
-.dcl.new-user-menu .email {
-  font-size: 14px;
-  line-height: 18px;
-}
-
-.dcl.new-user-menu .actions {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-  font-size: 14px;
-  border-top: 0.5px solid var(--clear-divider);
-  display: flex;
-  flex-direction: column;
-  gap: 1px;
-  padding: 16px 0px;
-}
-
-.dcl.new-user-menu .actions li,
-.dcl.new-user-menu .actions .item {
-  cursor: pointer;
-  padding: 0px 16px;
-  display: flex;
-  align-items: center;
-}
-
-.dcl.new-user-menu .actions li:hover,
-.dcl.new-user-menu .actions .item:hover {
-  background-color: var(--dropdown-hover);
-}
-
-.dcl.new-user-menu .actions li img {
-  margin-right: 12px;
-  display: block;
-  filter: var(--brightness);
-}
-
-.dcl.new-user-menu .actions .sign-out-icon {
-  width: 12px;
-  height: 12px;
-  filter: var(--brightness);
-  background: url(../../assets/signout.svg);
-  background-position: center;
-  background-repeat: no-repeat;
-  margin-right: 12px;
-}
-
-.dcl.new-user-menu .toggle {
-  width: 46px;
-  height: 46px;
-}
-
-.dcl.new-user-menu .menu .item,
-.dcl.new-user-menu .menu a {
-  color: var(--text);
-  font-weight: normal;
-}
-
-.dcl.new-user-menu .dcl.mana .symbol {
-  font-weight: bold;
-  color: var(--primary);
-}
-
-.dcl.new-user-menu-wrapper .dcl.mana {
-  cursor: pointer;
-}
-
-@media (max-width: 768px) {
-  .dcl.new-user-menu {
-    gap: 5px;
-    align-items: center;
-    flex-wrap: wrap;
-    justify-content: end;
-    margin-top: 15px;
-  }
-
-  .dcl.navbar .dcl.new-account-wrapper {
-    flex-direction: row;
-    justify-content: end;
-    order: 1;
-    width: 100%;
-  }
-
-  .dcl.navbar .dcl.new-account-wrapper .dcl.mana + .dcl.mana {
-    padding: 0;
-  }
-
-  .dcl.navbar:has(.dcl.mobile-user-balances-wrapper) {
-    height: calc(var(--navbar-height) + 16px);
-  }
-
-  .dcl.navbar:has(.dcl.mobile-user-balances-wrapper) > .ui.container {
-    align-items: flex-start;
-    padding-top: 8px;
-  }
-
-  .dcl.navbar:has(.dcl.mobile-user-balances-wrapper) .dcl.new-account-wrapper {
-    flex-direction: row;
-  }
-
-  .dcl.navbar:has(.dcl.mobile-user-balances-wrapper)
-    .dcl.new-account-wrapper
-    .dcl.mana {
-    margin-right: 8px;
-    padding: 0;
-  }
-
-  .dcl.navbar:has(.dcl.mobile-user-balances-wrapper)
-    .dcl.new-user-menu
-    .dcl.new-account-wrapper {
-    display: none;
-  }
-
-  .dcl.navbar:has(.dcl.mobile-user-balances-wrapper)
-    .dcl.new-account-wrapper
-    .dcl.mana:last-of-type {
-    margin-right: 0;
-  }
-
-  .dcl.navbar .dcl.mobile-user-balances-wrapper {
-    position: absolute;
-    top: 38px;
-    width: 240px;
-  }
-
-  .dcl.new-user-menu .toggle {
-    width: 36px;
-    height: 36px;
-  }
-
-  .dcl.new-user-menu .toggle .avatar-face {
-    width: 36px;
-    height: 36px;
-  }
-}
-
-.navbar-sign-in.ui.button {
-  min-width: unset;
-  max-width: 121px;
-  height: 34px;
-  display: flex;
-  align-items: center;
-}
-
-.dcl.new-user-menu .activity-icon {
-  padding: 0px !important;
-  margin: 0;
-}
-
-.dcl.new-account-wrapper {
-  display: flex;
-  gap: 15px;
-  align-items: center;
-}
-
-.dcl.new-user-menu .view-profile {
-  align-self: center;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 0px 16px 16px 16px;
-  width: 238px;
-}
-
-.dcl.new-user-menu .menu-option {
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-  height: 35px;
-}
