+++ src/index.ts
@@ -85,7 +85,6 @@ export * from './components/MultiStep/MultiStep'
 export * from './components/AuthorizationModal'
 export * from './components/CatalogCard/CatalogCard'
 export * from './components/AssetImage/AssetImage'
-export * from './components/UserInformationContainer/UserInformationContainer'
 export * from './components/ArrayFilter'
 export * from './components/InfoTooltip'
 export * from './components/RarityFilter'
